Added win/lose response. Added convenience buttons.
The game now properly tells you when the game is over and
if you won or lost.
Added in a button that will reset the game.
Added a button for instructions on how to play the game (it just
opens a link to the wiki page for Mastermind).
Added a checkbox for whether or not to show the winning circles.