Cleanup, commenting, fixed feedback logic.
Cleaned up some more portions of code.
Guess circles no longer get an onClick if they are not in
the top row or the game is over.
Added comments to the beginning of every method.
Fixed that the feedback circles were misaligned from the row
they should be associated with.